other(bump): bump vllm to 0.22.0 with optimum (#671)

+++ vllm_rbln/model_executor/models/optimum/qwen_vl.py
@@ -302,7 +302,8 @@ class RBLNOptimumQwenVLForConditionalGeneration(
                 " expected {padded_batch_size}.",
             )
 
-        inputs_embeds = self.model.embed_tokens(input_ids)
+        # NOTE: preprocess_prefill also use dtype casting.
+        inputs_embeds = self.model.embed_tokens(input_ids).to(self.dtype)
         position_embeds = []
         for b_id, request_id in enumerate(running_requests_ids):
             delta = cache_position[b_id] + self.rope_deltas[request_id]
